refactor: moves contexts settings into their own component
