+++ src/routes/(app)/settings/+page.svelte
@@ -1,18 +1,11 @@
 <script lang="ts">
-    import { type Context } from '@prisma/client';
 	import dayjs from "dayjs";
     import { user } from "$lib/stores/user";
-	import { contexts } from "$lib/stores/contexts";
 	import Link from "$lib/components/Link.svelte";
-	import TextInput from "$lib/components/TextInput.svelte";
-	import Button from "$lib/components/Button.svelte";
-	import { ContextRole } from "../../../types/contexts";
-	import Icon from '$lib/components/Icon.svelte';
 	import UpdateUserInfo from '$lib/components/forms/UpdateUserInfo.svelte';
 	import ChangePassword from '$lib/components/forms/ChangePassword.svelte';
-	import ContextModal from '$lib/components/modals/ContextModal.svelte';
+	import ContextSettings from '$lib/components/ContextSettings.svelte';
 
-    let contextsWithoutRoles: Context[] = [];
     let lastUpdated = '';
 
     $: {
@@ -32,10 +25,6 @@
             lastUpdated = `Your account information was last updated ${daysAgo} days ago.`;
         }
     }
-
-    $: if ($contexts && $contexts.length) {
-            contextsWithoutRoles = $contexts.filter(c => c.role === ContextRole.NONE);
-        }
 </script>
 
 <div class="settings-container no-scrollbar">
@@ -69,51 +58,7 @@
             <section data-testid="contexts-section" class="contexts-section">
                 <div data-testid="contexts-header" class="h4">Contexts</div>
                 <p>Manager the contexts that matter to you.</p>
-                <div data-testid="contexts" class="contexts-container">
-                    {#if $contexts && $contexts.length > 0}
-                        {#each contextsWithoutRoles as context}
-                            <div data-testid="{context.id}" class="context">
-                                <div class="context-info">
-                                    <div data-testid="context-name" class="context-name">
-                                        {context.name}
-                                    </div>
-                                    <p data-testid="context-description" class="context-description">{context.description}</p>
-                                </div>
-                                <div class="context-actions">
-                                    <Button
-                                        kind="neutral"
-                                        data-testid="edit-context-button"
-                                    >
-                                        Edit
-                                    </Button>
-    
-                                    <Button
-                                        kind="danger-transparent"
-                                        data-testid="delete-context-button"
-                                    >
-                                        <Icon name="trash" />
-                                    </Button>
-                                </div>
-                            </div>
-                        {/each}
-                    {:else}
-                        <div data-testid="no-contexts" class="no-contexts">
-                            <span>No contexts found</span>
-                        </div>
-                    {/if}
-                </div>
-    
-                <div class="section-actions">
-                    <ContextModal let:openContextModal>
-                        <Button
-                            kind="neutral"
-                            data-testid="add-context-button"
-                            on:click={() => openContextModal()}
-                        >
-                            + Add Context
-                        </Button>
-                    </ContextModal>
-                </div>
+                <ContextSettings />
             </section>
         </div>
 
@@ -211,17 +156,6 @@
             color: var(--dark-900);
         }
 
-        & form {
-            display: flex;
-            flex-direction: column;
-            flex-grow: 1;
-            gap: 1rem;
-
-            &.user-info-container-form {
-                justify-content: space-between;
-            }
-        }
-
         & footer {
             display: flex;
             justify-content: space-between;
@@ -229,12 +163,6 @@
             gap: 1rem;
             margin-top: 1rem;
         }
-    }
-
-    .current-password-inputs-container {
-        display: flex;
-        flex-direction: column;
-        gap: 0.5rem;
     }
 
     .section-actions {
@@ -248,56 +176,6 @@
     .user-info-section {
         container-type: inline-size;
         container-name: user-info-section;
-    }
-
-    .user-info-container {
-        display: grid;
-        grid-template-columns: 1fr;
-        gap: 0.5rem;
-    }
-
-    .user-info {
-        & span:first-child {
-            margin-right: 0.5rem;
-            color: var(--secondary-500);
-        }
-    }
-
-    .contexts-container {
-        display: grid;
-        grid-template-columns: 1fr;
-        padding-bottom: 0.75rem;
-    }
-
-    .context {
-        display: flex;
-        justify-content: space-between;
-        padding: 0.5rem;
-        border-bottom: 1px solid var(--dark-400);
-    }
-
-    .context-info {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        gap: 0.3rem;
-    }
-
-    .context-name {
-        font-size: 1rem;
-    }
-
-    .context-description {
-        font-size: 0.875rem;
-        color: var(--dark-900);
-    }
-
-    .context-actions {
-        display: flex;
-        justify-content: flex-end;
-        align-items: center;
-        gap: 0.5rem;
-        width: 10rem;
     }
 
     .metadata-container {
@@ -337,12 +215,6 @@
 
         .contexts-section {
             grid-column: span 2;
-        }
-    }
-
-    @container user-info-section (min-width: 600px) {
-        .user-info-container {
-            grid-template-columns: 1fr 1fr;
         }
     }
 </style>
